Wikibot.recentChanges(): always convert Calendar params to UTC time zone
This method was fed with Calendar variables that could refer to any time
zone possible. However, the timestamp string built internally and passed
to the URL request assumes UTC everywhere, so some standardization is
required. This was failing on plwiktionary-related scripts when calling
Wiki.makeCalendar() (the default time zone set here is 'Europe/Warsaw').